+++ src/components/index.js
@@ -24,7 +24,6 @@ import FakeSigninForm from './forms/FakeSigninForm'
 import FileInput from './forms/FileInput'
 import Input from './forms/Input'
 import Select from './forms/Select'
-import SteamSignin from './forms/SteamSignin'
 
 // Icons
 import AcceptIcon from './icons/AcceptIcon'
@@ -68,7 +67,6 @@ import UserIcon from './icons/UserIcon'
 import WarningCircleIcon from './icons/WarningCircleIcon'
 
 // Lobby
-import LobbyHeader from './lobby/LobbyHeader'
 import LobbyLineup from './lobby/LobbyLineup'
 import LobbyModeSelector from './lobby/LobbyModeSelector'
 import LobbyPlayButton from './lobby/LobbyPlayButton'
@@ -77,11 +75,7 @@ import LobbySeat from './lobby/LobbySeat'
 import MatchFoundModal from './lobby/MatchFoundModal'
 
 // User
-import FriendListUser from './user/FriendListUser'
 import LevelBadge from './user/LevelBadge'
-import UserCard from './user/UserCard'
-import UserCardMini from './user/UserCardMini'
-import UserStatus from './user/UserStatus'
 
 // Match
 import LevelProgressBar from './match/LevelProgressBar'
@@ -151,7 +145,6 @@ export {
   FriendList,
   FriendListGroup,
   FriendListGroupItem,
-  FriendListUser,
   FriendsIcon,
   GroupAddIcon,
   HeaderProfile,
@@ -166,7 +159,6 @@ export {
   LevelStatsCard,
   Loading,
   LoadingBackdrop,
-  LobbyHeader,
   LobbyLineup,
   LobbyModeSelector,
   LobbyPlayButton,
@@ -200,7 +192,6 @@ export {
   ShopIcon,
   Sidebar,
   StarIcon,
-  SteamSignin,
   StoreIcon,
   SupportIcon,
   SupportModal,
@@ -208,9 +199,6 @@ export {
   ToastList,
   ToastListItem,
   TrashIcon,
-  UserCard,
-  UserCardMini,
   UserIcon,
-  UserStatus,
   WarningCircleIcon,
 }